hopefully fix the missing fefe issue Problem
Depending on the dst and src port, scapy sometimes detects a NetflowHeader field inside the UDP packet and strips two bytes for the short version. So we can't use .load to get the innermost packet - that sometimes misses it's fefe. Instead get whatever is inside the UDP packet (either NetflowHeader or Raw) and use it's original bytes.